rewrite xUnit tests
